updated screenBrightness

+++ Assets/MXR.SDK/Runtime/Types/StatusTypes.cs
@@ -60,7 +60,7 @@ namespace MXR.SDK {
         /// <summary>
         /// The brightness level for the screen
         /// </summary>
-        public int screenBrightness;
+        public float screenBrightness;
         
         /// <summary>
         /// The last app open in the foregound that is appropriate to show in the shortcut menu.
